requirements.txt
keyboard
mss
Pillow
pytesseract
googletrans==4.0.0-rc1